[dy] Refactor logs
