+++ docker-compose.yml
@@ -13,7 +13,6 @@ services:
       - ECS_TASK_DEFINITION=$ECS_TASK_DEFINITION
       - ECS_CONTAINER_NAME=$ECS_CONTAINER_NAME
       - MAGE_DATABASE_CONNECTION_URL=$DATABASE_CONNECTION_URL
-      - MAGE_LOG_STORAGE_LOCATION=$LOG_STORAGE_LOCATION
     ports:
       - 6789:6789
     volumes:
