+++ mage_ai/data_preparation/executors/block_executor.py
@@ -1,5 +1,5 @@
-from mage_ai.data_preparation.logger_manager import LoggerManager
 from mage_ai.data_preparation.logging.logger import DictLogger
+from mage_ai.data_preparation.logging.logger_manager_factory import LoggerManagerFactory
 from mage_ai.shared.hash import merge_dict
 from typing import Callable, Dict
 import json
@@ -12,13 +12,13 @@ class BlockExecutor:
         self.block_uuid = block_uuid
         self.block = self.pipeline.get_block(block_uuid)
         self.execution_partition = execution_partition
-        self.logger_manager = LoggerManager(
+        self.logger_manager = LoggerManagerFactory.get_logger_manager(
             pipeline_uuid=self.pipeline.uuid,
             block_uuid=self.block_uuid,
             partition=self.execution_partition,
             repo_config=self.pipeline.repo_config,
         )
-        self.logger = DictLogger(self.logger_manager.get_logger())
+        self.logger = DictLogger(self.logger_manager.logger)
 
     def execute(
         self,
