+++ mage_ai/data_preparation/executors/pyspark_block_executor.py
@@ -1,7 +1,7 @@
 from contextlib import redirect_stdout
 from mage_ai.data_preparation.executors.block_executor import BlockExecutor
-from mage_ai.data_preparation.logger_manager import StreamToLogger
 from mage_ai.data_preparation.models.pipeline import Pipeline
+from mage_ai.data_preparation.shared.stream import StreamToLogger
 from mage_ai.data_preparation.templates.utils import template_env
 from mage_ai.services.aws.emr import emr
 from mage_ai.services.aws.emr.resource_manager import EmrResourceManager
