+++ mage_ai/data_preparation/executors/streaming_pipeline_executor.py
@@ -1,8 +1,8 @@
 from contextlib import redirect_stdout
 from mage_ai.data_preparation.executors.pipeline_executor import PipelineExecutor
-from mage_ai.data_preparation.logger_manager import StreamToLogger
 from mage_ai.data_preparation.models.constants import BlockType
 from mage_ai.data_preparation.models.pipeline import Pipeline
+from mage_ai.data_preparation.shared.stream import StreamToLogger
 from typing import Callable, Dict
 import yaml
