--- mage_ai/data_preparation/logging/s3/config.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-from mage_ai.shared.config import BaseConfig
-
-
-@dataclass
-class S3Config(BaseConfig):
-    bucket: str
-    prefix: str
