+++ mage_ai/data_preparation/shared/constants.py
@@ -1,4 +1,2 @@
 REPO_PATH_ENV_VAR = 'MAGE_REPO_PATH'
 MANAGE_ENV_VAR = 'MAGE_IS_MANAGE_INSTANCE'
-LOG_STORAGE_LOCATION_ENV_VAR = 'MAGE_LOG_STORAGE_LOCATION'
-LOG_LEVEL_ENV_VAR = 'MAGE_LOG_LEVEL'
